src/utils/converters.py
"""
File conversion utilities for MarkItDown Web
"""
from pathlib import Path
from typing import Optional, Dict, Any
import mimetypes
import tempfile

class ConversionError(Exception):
    """Custom exception for conversion errors"""
    pass

def detect_file_type(file_path: Path) -> str:
    """
    Detect the type of file based on its extension and mime type
    
    Args:
        file_path: Path to the file
        
    Returns:
        String indicating the file type
    """
    mime_type, _ = mimetypes.guess_type(str(file_path))
    if mime_type:
        return mime_type
    return "application/octet-stream"

def convert_to_markdown(
    file_path: Path,
    options: Optional[Dict[str, Any]] = None
) -> str:
    """
    Convert a file to Markdown format
    
    Args:
        file_path: Path to the input file
        options: Dictionary of conversion options
        
    Returns:
        Markdown string
        
    Raises:
        ConversionError: If conversion fails
    """
    try:
        # TODO: Implement actual conversion logic using MarkItDown library
        file_type = detect_file_type(file_path)
        
        # Placeholder implementation
        return f"""# Converted from {file_path.name}

File type: {file_type}

Content will be converted here.
"""
    except Exception as e:
        raise ConversionError(f"Failed to convert {file_path}: {e}")

def extract_metadata(file_path: Path) -> Dict[str, Any]:
    """
    Extract metadata from a file
    
    Args:
        file_path: Path to the file
        
    Returns:
        Dictionary of metadata
    """
    try:
        # TODO: Implement metadata extraction
        return {
            "filename": file_path.name,
            "size": file_path.stat().st_size,
            "type": detect_file_type(file_path)
        }
    except Exception as e:
        print(f"Error extracting metadata: {e}")
        return {} 